ads-bare-metal/src/requirements.txt
google-auth>=2.9.0
google-api-core>=2.8.2
google-cloud-bigquery>=3.2.0
google-cloud-storage>=2.4.0
google-ads>=20.0.0
proto-plus>=1.19.6
protobuf>=3.20.0
google-cloud-firestore>=2.5.0
function-flow>=0.0.7
gspread>=5.3.2
oauth2client
